Added substring method
