+++ src/main/scala/scala/slick/driver/SQLiteDriver.scala
@@ -94,6 +94,7 @@ trait SQLiteDriver extends JdbcDriver { driver =>
     override def expr(c: Node, skipParens: Boolean = false): Unit = c match {
       case Library.UCase(ch) => b"upper(!$ch)"
       case Library.LCase(ch) => b"lower(!$ch)"
+      case Library.Substring(n, start, end) => b"\(substr($n,$start,$end)\)"
       case Library.%(l, r) => b"\($l%$r\)"
       case Library.Ceiling(ch) => b"round($ch+0.5)"
       case Library.Floor(ch) => b"round($ch-0.5)"
